swap updates
