+++ Frontend-v1-Original/components/ssSwap/select.tsx
@@ -1,0 +1,313 @@
+import { useEffect, useState } from "react";
+import { isAddress } from "viem";
+import {
+  Button,
+  Dialog,
+  IconButton,
+  InputAdornment,
+  MenuItem,
+  TextField,
+  Typography,
+} from "@mui/material";
+import { DeleteOutline, Search } from "@mui/icons-material";
+import BigNumber from "bignumber.js";
+
+import { BaseAsset } from "../../stores/types/types";
+import stores from "../../stores";
+import { ACTIONS, ETHERSCAN_URL } from "../../stores/constants/constants";
+import { formatCurrency } from "../../utils/utils";
+
+export function AssetSelect({
+  type,
+  value,
+  assetOptions,
+  onSelect,
+}: {
+  type: string;
+  value: BaseAsset | null;
+  assetOptions: BaseAsset[] | undefined;
+  onSelect: (_type: string, _asset: BaseAsset) => void;
+}) {
+  const [open, setOpen] = useState(false);
+  const [search, setSearch] = useState("");
+  const [filteredAssetOptions, setFilteredAssetOptions] =
+    useState<BaseAsset[]>();
+
+  const [manageLocal, setManageLocal] = useState(false);
+
+  const openSearch = () => {
+    setSearch("");
+    setOpen(true);
+  };
+
+  useEffect(() => {
+    async function sync() {
+      let ao = assetOptions?.filter((asset) => {
+        if (search && search !== "") {
+          return (
+            asset.address.toLowerCase().includes(search.toLowerCase()) ||
+            asset.symbol.toLowerCase().includes(search.toLowerCase()) ||
+            asset.name.toLowerCase().includes(search.toLowerCase())
+          );
+        } else {
+          return true;
+        }
+      });
+
+      setFilteredAssetOptions(ao);
+
+      //no options in our default list and its an address we search for the address
+      if (
+        ao &&
+        ao.length === 0 &&
+        search &&
+        search.length === 42 &&
+        isAddress(search)
+      ) {
+        const baseAsset = await stores.stableSwapStore.getBaseAsset(
+          search,
+          true,
+          true
+        );
+        if (baseAsset) {
+          stores.emitter.emit(ACTIONS.WARNING, {
+            warning: "Token is not whitelisted",
+          });
+        }
+      }
+    }
+    sync();
+    return () => {};
+  }, [assetOptions, search]);
+
+  const onSearchChanged = (event: React.ChangeEvent<HTMLInputElement>) => {
+    setSearch(event.target.value);
+  };
+
+  const onLocalSelect = (type: string, asset: BaseAsset) => {
+    setSearch("");
+    setManageLocal(false);
+    setOpen(false);
+    onSelect(type, asset);
+  };
+
+  const onClose = () => {
+    setManageLocal(false);
+    setSearch("");
+    setOpen(false);
+  };
+
+  const toggleLocal = () => {
+    setManageLocal(!manageLocal);
+  };
+
+  const deleteOption = (token: BaseAsset) => {
+    stores.stableSwapStore.removeBaseAsset(token);
+  };
+
+  const viewOption = (token: BaseAsset) => {
+    window.open(`${ETHERSCAN_URL}token/${token.address}`, "_blank");
+  };
+
+  const renderManageOption = (type: string, asset: BaseAsset, idx: number) => {
+    return (
+      <MenuItem
+        defaultValue={asset.address}
+        key={asset.address + "_" + idx}
+        className="flex items-center justify-between px-0"
+      >
+        <div className="relative mr-3 w-14">
+          <img
+            className="h-full w-full rounded-[30px] border border-[rgba(126,153,153,0.5)] bg-[rgb(33,43,72)] p-[6px]"
+            alt=""
+            src={asset ? `${asset.logoURI}` : ""}
+            height="60px"
+            onError={(e) => {
+              (e.target as HTMLImageElement).onerror = null;
+              (e.target as HTMLImageElement).src = "/tokens/unknown-logo.png";
+            }}
+          />
+        </div>
+
+        <div>
+          <Typography variant="h5">{asset ? asset.symbol : ""}</Typography>
+          <Typography variant="subtitle1" color="textSecondary">
+            {asset ? asset.name : ""}
+          </Typography>
+        </div>
+        <div className="flex flex-[1] justify-end">
+          <IconButton
+            onClick={() => {
+              deleteOption(asset);
+            }}
+          >
+            <DeleteOutline />
+          </IconButton>
+          <IconButton
+            onClick={() => {
+              viewOption(asset);
+            }}
+          >
+            ↗
+          </IconButton>
+        </div>
+      </MenuItem>
+    );
+  };
+
+  const renderAssetOption = (type: string, asset: BaseAsset, idx: number) => {
+    return (
+      <MenuItem
+        defaultValue={asset.address}
+        key={asset.address + "_" + idx}
+        className="flex items-center justify-between px-0"
+        onClick={() => {
+          onLocalSelect(type, asset);
+        }}
+      >
+        <div className="relative mr-3 w-14">
+          <img
+            className="h-full w-full rounded-[30px] border border-[rgba(126,153,153,0.5)] bg-[rgb(33,43,72)] p-[6px]"
+            alt=""
+            src={asset ? `${asset.logoURI}` : ""}
+            height="60px"
+            onError={(e) => {
+              (e.target as HTMLImageElement).onerror = null;
+              (e.target as HTMLImageElement).src = "/tokens/unknown-logo.png";
+            }}
+          />
+        </div>
+        <div>
+          <Typography variant="h5">{asset ? asset.symbol : ""}</Typography>
+          <Typography variant="subtitle1" color="textSecondary">
+            {asset ? asset.name : ""}
+          </Typography>
+        </div>
+        <div className="ml-12 flex flex-[1] flex-col items-end">
+          <Typography variant="h5">
+            {asset && asset.balance ? formatCurrency(asset.balance) : "0.00"}
+          </Typography>
+          <Typography variant="subtitle1" color="textSecondary">
+            Balance
+          </Typography>
+        </div>
+      </MenuItem>
+    );
+  };
+
+  const renderManageLocal = () => {
+    return (
+      <>
+        <div className="h-[600px] overflow-y-scroll p-6">
+          <TextField
+            autoFocus
+            variant="outlined"
+            fullWidth
+            placeholder="CANTO, NOTE, 0x..."
+            value={search}
+            onChange={onSearchChanged}
+            InputProps={{
+              startAdornment: (
+                <InputAdornment position="start">
+                  <Search />
+                </InputAdornment>
+              ),
+            }}
+          />
+
+          <div className="mt-3 flex w-full min-w-[390px] flex-col">
+            {filteredAssetOptions
+              ? filteredAssetOptions
+                  .filter((option) => {
+                    return option.local === true;
+                  })
+                  .map((asset, idx) => {
+                    return renderManageOption(type, asset, idx);
+                  })
+              : []}
+          </div>
+        </div>
+        <div className="flex w-full items-center justify-center p-[6px]">
+          <Button onClick={toggleLocal}>Back to Assets</Button>
+        </div>
+      </>
+    );
+  };
+
+  const renderOptions = () => {
+    return (
+      <>
+        <div className="h-[600px] overflow-y-scroll p-6">
+          <TextField
+            autoFocus
+            variant="outlined"
+            fullWidth
+            placeholder="CANTO, NOTE, 0x..."
+            value={search}
+            onChange={onSearchChanged}
+            InputProps={{
+              startAdornment: (
+                <InputAdornment position="start">
+                  <Search />
+                </InputAdornment>
+              ),
+            }}
+          />
+
+          <div className="mt-3 flex w-full min-w-[390px] flex-col">
+            {filteredAssetOptions
+              ? filteredAssetOptions
+                  .sort((a, b) => {
+                    if (BigNumber(a.balance || 0).lt(b.balance || 0)) return 1;
+                    if (BigNumber(a.balance || 0).gt(b.balance || 0)) return -1;
+                    if (a.symbol.toLowerCase() < b.symbol.toLowerCase())
+                      return -1;
+                    if (a.symbol.toLowerCase() > b.symbol.toLowerCase())
+                      return 1;
+                    return 0;
+                  })
+                  .map((asset, idx) => {
+                    return renderAssetOption(type, asset, idx);
+                  })
+              : []}
+          </div>
+        </div>
+        <div className="flex w-full items-center justify-center p-[6px]">
+          <Button onClick={toggleLocal}>Manage Local Assets</Button>
+        </div>
+      </>
+    );
+  };
+
+  return (
+    <>
+      <div
+        className="min-h-[100px] p-3"
+        onClick={() => {
+          openSearch();
+        }}
+      >
+        <div className="relative w-full cursor-pointer">
+          <img
+            className="h-full w-full rounded-[50px] border border-[rgba(126,153,153,0.5)] bg-[#032725] p-[10px]"
+            alt=""
+            src={value ? `${value.logoURI}` : ""}
+            height="100px"
+            onError={(e) => {
+              (e.target as HTMLImageElement).onerror = null;
+              (e.target as HTMLImageElement).src = "/tokens/unknown-logo.png";
+            }}
+          />
+        </div>
+      </div>
+      <Dialog
+        onClose={onClose}
+        aria-labelledby="asset-select-dialog-title"
+        open={open}
+      >
+        {!manageLocal && renderOptions()}
+        {manageLocal && renderManageLocal()}
+      </Dialog>
+    </>
+  );
+}
