--- Frontend-v1-Original/utils/package.json
@@ -1,3 +1,0 @@
-{
-  "main": "utils.js"
-}
